move prep robots to constants folder
